requirements.txt
fluidsynth==0.2
matplotlib==3.10.0
mido==1.3.3
numpy==2.1.3
pandas==2.2.3
PyQt6==6.8.1
PyYAML==6.0.2
seaborn==0.13.2
torch==2.5.1
